scripts/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.2
python-multipart==0.0.6
aiofiles==23.2.1
